+++ CHANGELOG.md
@@ -110,3 +110,7 @@
 * Define `Different` type alias
 * Define `Matching` type alias
 * Define `Record` type family
+* Define `Construction` type alias
+* Define `Instruction` type alias
+* Define `Instruct` patter synonym
+* Define `Load` patter synonym